DB: bestellungLoeschen aktualisiert.

+++ FinalProject/src/model/DB.java
@@ -910,7 +910,7 @@ public class DB implements Configuration {
 	 */
 	public void bestellungSpeichern(String bstid, String bestelltext, String anleger, String status,
 									String bestelltermin, String kid, String[][] bpos) throws SQLException, NotExistInDatabaseException {
-		if ( !checkIfElementExists(TABLE_OWNER + "." + TABLE_KUNDE, "kid", kid) ) {
+		if ( !checkIfElementExists(TABLE_KUNDE, "kid", kid) ) {
 			throw new NotExistInDatabaseException("Kunde mit der ID " + kid + " existiert nicht in der Datenbank.");
 		}
 		this.connection.setAutoCommit(false);
@@ -1032,7 +1032,7 @@ public class DB implements Configuration {
 	public boolean bestellungAendern(String bstid, String bestelltext, String anleger,
 									 String bestelltermin, String kid, String[][] bpos) throws SQLException, NotExistInDatabaseException{
 
-		if ( !checkIfElementExists(TABLE_OWNER + "." + TABLE_KUNDE, "kid", kid) ) {
+		if ( !checkIfElementExists(TABLE_KUNDE, "kid", kid) ) {
 			throw new NotExistInDatabaseException("Kunde mit der ID " + kid + " existiert nicht in der Datenbank.");
 		}
 
@@ -1212,7 +1212,6 @@ public class DB implements Configuration {
 		return success;
 	}
 
-
 	public boolean deleteRow(String table, String condition) throws SQLException {
 		this.connection.setAutoCommit(false);
 		Savepoint savePoint1 = this.connection.setSavepoint();
@@ -1241,6 +1240,14 @@ public class DB implements Configuration {
 		return success;
 	}
 
+	/**
+	 *
+	 * @param bstid
+	 * @return
+	 * @throws NotExistInDatabaseException
+	 * @throws SQLException
+	 * @throws ParseException
+	 */
 	public boolean bestellungLoeschen(String bstid) throws NotExistInDatabaseException, SQLException, ParseException {
 		this.connection.setAutoCommit(false);
 		Statement stmt = null;
@@ -1251,11 +1258,19 @@ public class DB implements Configuration {
 
 		try {
 			stmt = this.connection.createStatement();
-//
-//			// Wenn die Bestellung noch OFFEN oder bereits ERLEDIGT ist, kann sie nicht mehr ausgeliefert werden.
-//			if ( status.trim().equals("OFFEN") || status.trim().equals("ERLEDIGT") ) {
-//				return false;
-//			}
+			sql_query = "SELECT status FROM " + TABLE_OWNER + ".BESTELLUNG " +
+						"WHERE bstid = " + bstid + " FOR UPDATE NOWAIT";
+			ResultSet rs = stmt.executeQuery(sql_query);
+			if ( !rs.next() ) {
+				success = false;
+				throw new NotExistInDatabaseException("<html>Invalid BSTID. Ung&uuml;ltige Bestellungs-ID.</html>");
+			}
+
+			String status = rs.getString("STATUS");
+			// Wenn die Bestellung ERLEDIGT ist, kann sie nicht geloescht werden.
+			if ( status.trim().equals("ERLEDIGT") ) {
+				return false;
+			}
 
 			// Loesche alle Bestellpositionen.
 			sql_query = "DELETE FROM " + TABLE_OWNER + ".BESTELLPOSITION " +
